Factored out ReifiedTerminalPlate
